+++ app/src/civ6IntroSkip.js
@@ -1,9 +1,9 @@
 import electron from "electron";
 import * as fs from "fs";
 import * as path from "path";
-import { execFileSync } from "child_process";
 import { default as log } from "electron-log";
 import { RPC_INVOKE } from "./rpcChannels.js";
+import { findGameInstallDir } from "./civ6Installation.js";
 import {
   findAppOptionsPath,
   getAppOption,
@@ -18,7 +18,6 @@ import {
 // For Windows we can go a step farther and prevent the game from opening the two logo
 // videos as well by holding open handles against them with exclusive read.
 
-const CIV6_STEAM_DIR = "Sid Meier's Civilization VI";
 const LOGO_MOVIES = ["logos.bk2", "LOGO_2KFiraxis.bk2"];
 
 const UV_FS_O_EXLOCK = 0x10000000; // Maps to FILE_SHARE_NONE on Windows
@@ -27,89 +26,6 @@ const LOGO_LOCK_START_TIMEOUT_MS = 5 * 60 * 1000;
 const LOGO_LOCK_MAX_MS = 6 * 60 * 60 * 1000;
 const REVERT_POLL_MS = 10 * 1000;
 const REVERT_MAX_MS = 6 * 60 * 60 * 1000;
-
-const regQuery = (key, value) => {
-  try {
-    const out = execFileSync("reg", ["query", key, "/v", value], { encoding: "utf8", windowsHide: true });
-    const m = new RegExp(`${value}\\s+REG_SZ\\s+(.+)`, "i").exec(out);
-
-    return m ? m[1].trim() : null;
-  } catch {
-    return null;
-  }
-};
-
-const steamLibraries = () => {
-  const root =
-    regQuery("HKCU\\Software\\Valve\\Steam", "SteamPath") ||
-    regQuery("HKLM\\SOFTWARE\\WOW6432Node\\Valve\\Steam", "InstallPath") ||
-    "C:\\Program Files (x86)\\Steam";
-  const libs = [root];
-
-  try {
-    const vdf = fs.readFileSync(path.join(root, "steamapps", "libraryfolders.vdf"), "utf8");
-
-    for (const m of vdf.matchAll(/"path"\s+"([^"]+)"/g)) {
-      libs.push(m[1].replace(/\\\\/g, "\\"));
-    }
-  } catch {
-    // No extra libraries
-  }
-
-  return libs;
-};
-
-const findSteamCiv6 = () =>
-  steamLibraries()
-    .map(lib => path.join(lib, "steamapps", "common", CIV6_STEAM_DIR))
-    .find(p => fs.existsSync(p)) || null;
-
-const findEpicCiv6 = () => {
-  const manifests = path.join(
-    process.env.ProgramData || "C:\\ProgramData",
-    "Epic",
-    "EpicGamesLauncher",
-    "Data",
-    "Manifests",
-  );
-
-  try {
-    for (const f of fs.readdirSync(manifests)) {
-      if (!f.endsWith(".item")) {
-        continue;
-      }
-
-      try {
-        const item = JSON.parse(fs.readFileSync(path.join(manifests, f), "utf8"));
-
-        if (
-          /civilization vi\b/i.test(item.DisplayName || "") &&
-          item.InstallLocation &&
-          fs.existsSync(item.InstallLocation)
-        ) {
-          return item.InstallLocation;
-        }
-      } catch {
-        // Malformed manifest
-      }
-    }
-  } catch {
-    // No Epic launcher
-  }
-
-  return null;
-};
-
-// Windows only, find the installation folder for Civ 6
-export const findGameInstallDir = dataPath => {
-  if (process.platform !== "win32") {
-    return null;
-  }
-
-  const epic = /\(Epic\)$/i.test(path.basename(dataPath));
-
-  return (epic ? findEpicCiv6() : findSteamCiv6()) || findSteamCiv6() || findEpicCiv6();
-};
 
 export const logoMoviePaths = dataPath => {
   const install = findGameInstallDir(dataPath);
